Use app router.
