+++ next.config.js
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   generateEtags: false,
+  experimental: {
+    appDir: true
+  }
 };
 
 module.exports = nextConfig;
